split guide code into separate files
