+++ docs/src/main/paradox/getting-started/projection-handler.md
@@ -16,11 +16,6 @@ artifact=akka-projection-cassandra_$scala.binary.version$
 version=$project.version$
 }
 
-Use the following package and imports in your project:
-
-Scala
-:  @@snip [ShoppingCartApp.scala](/examples/src/test/scala/docs/guide/ShoppingCartApp.scala) { #guideProjectionImports }
-
 It's the user's responsibility to implement the means to project into the target system, the Projection itself will only manage the persistence of the offset (though it is possible to enlist your projection into transactions when using projection implementations that support exactly-once like the @ref:[JDBC](../jdbc.md)).
 This guide encapsulates its data access layer in a Repository called the `ItemPopularityProjectionRepository`.
 The repository will manage a Cassandra table called `item_popularity`.
@@ -34,11 +29,15 @@ However, since the count is only a rough metric to judge how popular an item is 
 
 @@@
 
+Add the `ItemPopularityProjectionRepository` to your project:
+
 Scala
-:  @@snip [ShoppingCartApp.scala](/examples/src/test/scala/docs/guide/ShoppingCartApp.scala) { #guideProjectionRepo }
+:  @@snip [ItemPopularityProjectionRepository.scala](/examples/src/test/scala/docs/guide/ItemPopularityProjectionRepository.scala) { #guideProjectionRepo }
 
 Now it's time to write the Projection handler itself.
 This example uses a @apidoc[Handler] that will process `ShoppingCartEvents.Event` events from the @apidoc[SourceProvider] that we implemented earlier.
+Specifically, it will only process `ItemEvents` that modify the items added or removed from a shopping cart.
+It will ignore all shopping cart `Checkout` events by skipping them.
 The event envelopes are processed in the `process` method. 
 
 This example will also log the popularity count of every 10th item event that is processed.
@@ -46,9 +45,16 @@ The logging counter is stored as a mutable variable within the handler.
 Since this is a simple log operation managing the state in this manner is fine, but to handle more advanced stateful operations you should evaluate using the @apidoc[StatefulHandler].
 
 Scala
-:  @@snip [ShoppingCartApp.scala](/examples/src/test/scala/docs/guide/ShoppingCartApp.scala) { #guideProjectionHandler }
+:  @@snip [ItemPopularityProjectionHandler.scala](/examples/src/test/scala/docs/guide/ItemPopularityProjectionHandler.scala) { #guideProjectionHandler }
 
 The projection is run by wrapping it in a @apidoc[ProjectionBehavior$] and spawning it as an Actor in the @apidoc[akka.actor.typed.ActorSystem].
+
+Add the following imports to `ShoppingCartApp`:
+
+Scala
+:  @@snip [ShoppingCartApp.scala](/examples/src/test/scala/docs/guide/ShoppingCartApp.scala) { #guideProjectionImports }
+
+Setup the Projection in the Guardian `Behavior` defined in `ShoppingCartApp`:
 
 Scala
 :  @@snip [ShoppingCartApp.scala](/examples/src/test/scala/docs/guide/ShoppingCartApp.scala) { #guideProjectionSetup }
