+++ examples/src/test/scala/docs/guide/ItemPopularityProjectionRepository.scala
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2020 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+//#guideProjectionRepo
+package docs.guide
+
+import scala.concurrent.ExecutionContext
+import scala.concurrent.Future
+
+import akka.Done
+import akka.stream.alpakka.cassandra.scaladsl.CassandraSession
+
+trait ItemPopularityProjectionRepository {
+
+  def update(itemId: String, delta: Int): Future[Done]
+  def getItem(itemId: String): Future[Option[Long]]
+}
+
+class ItemPopularityProjectionRepositoryImpl(session: CassandraSession)(implicit val ec: ExecutionContext)
+    extends ItemPopularityProjectionRepository {
+
+  val keyspace = "akka_projection"
+  val popularityTable = "item_popularity"
+
+  override def update(itemId: String, delta: Int): Future[Done] = {
+    session.executeWrite(
+      s"UPDATE $keyspace.$popularityTable SET count = count + ? WHERE item_id = ?",
+      java.lang.Long.valueOf(delta),
+      itemId)
+  }
+
+  override def getItem(itemId: String): Future[Option[Long]] = {
+    session
+      .selectOne(s"SELECT item_id, count FROM $keyspace.$popularityTable WHERE item_id = ?", itemId)
+      .map(opt => opt.map(row => row.getLong("count").longValue()))
+  }
+}
+//#guideProjectionRepo
