+++ examples/src/test/scala/docs/guide/ShoppingCartTags.scala
@@ -1,0 +1,12 @@
+/*
+ * Copyright (C) 2020 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+//#guideTags
+package docs.guide
+
+object ShoppingCartTags {
+  val single = "shopping-cart"
+  val tags = Vector("carts-0", "carts-1", "carts-2")
+}
+//#guideTags
